backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
alembic>=1.13.1
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
asyncpg>=0.29.0
greenlet>=3.0.3
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.125.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0
pytest>=8.1.0
pytest-asyncio>=0.23.6
